requirements.txt
flask
gunicorn
google-generativeai
python-dotenv
numpy
pandas